% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgEvalLogLast.R
\name{avgEvalLogLast}
\alias{avgEvalLogLast}
\title{avgEvalLogLast function}
\usage{
avgEvalLogLast(path, top, mutation, initJar, rp, nround)
}
\arguments{
\item{path}{results path}

\item{top}{number of top ranks}

\item{mutation}{name of mutation}

\item{initJar}{t of f}

\item{rp}{number of repetitions}

\item{nround}{number of rounds

avgEvalLogLast(path, top, mutation, initJar, rp, nround)}
}
\description{
This function calculates train and test error avgerage from log file
}
