% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{check_dir() function}
\usage{
check_dir(mainDir, subDir)
}
\arguments{
\item{mainDir}{sourcePath}

\item{subDir}{mutation_rawdata name}
}
\description{
This function checks mainDir and subdir
If there is no directory, create one.
}
\keyword{:}
\keyword{and}
\keyword{check}
\keyword{create}
\keyword{dir}
