% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COOBoost.R, R/COOBoostR.R
\docType{package}
\name{COOBoostR}
\alias{COOBoostR}
\title{COOBoostR}
\usage{
COOBoostR(
  sourcePath,
  resultPath,
  epimarker_rawdata,
  mutation_rawdata,
  mEta = 0.03,
  mdepth = 2
)
}
\arguments{
\item{sourcePath}{sourcePath, must end with a dir delimiter(ex : /)}

\item{resultPath}{resultPath, must end with a dir delimiter(ex : /)}

\item{epimarker_rawdata}{epimarker data}

\item{mutation_rawdata}{mutation data}

\item{mEta}{Learning Rate (default 0.03)}

\item{mdepth}{max depth (default 2)
COOBoost_main(path, top, mutation, initJar, rp, nround)}
}
\description{
An extreme gradient boosting-based tool for robust tissue or cell-of-origin prediction of tumors

This function finds Cell Of Origin by xgboost
}
