% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutTo1mb.R
\name{MutTo1mb}
\alias{MutTo1mb}
\title{MutTo1mb() function}
\usage{
MutTo1mb(
  bedpath,
  resultpath,
  mb1_path,
  bed_sep = ",",
  bed_header = F,
  bedform = "bed4"
)
}
\arguments{
\item{bedpath}{bedfile path}

\item{resultpath}{results path}

\item{mb1_path}{mb1 file path}

\item{bed_sep}{bed file separator, default = ","}

\item{bed_header}{bed file header, default = F

EpiTo1mb(bedpath, resultpath, mb1_path, bedform, bed_sep, bed_header)}

\item{bedform}{bed4 or bed5, default = "bed4"}
}
\description{
This function preprocesses the bed file to fit 1mb.
}
