% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_BEDs.R
\name{Prep_BEDs}
\alias{Prep_BEDs}
\title{Prep_BEDs() function}
\usage{
Prep_BEDs(
  BED_path,
  resultpath,
  mb1_path,
  BED_sep = "\\t",
  BED_header = F,
  BED_form = "bed4"
)
}
\arguments{
\item{BED_path}{bedfile path}

\item{resultpath}{results path}

\item{mb1_path}{mb1 file path}

\item{BED_sep}{bed file separator, default = tap}

\item{BED_header}{bed file header, default = F

Prep_BEDs(BED_path, resultpath, mb1_path, BED_form, BED_sep, BED_header)}

\item{BED_form}{bed4 or bed5, default = "bed4"}
}
\description{
This function preprocesses the bed file to fit 1mb.
}
