% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{readFile() function}
\usage{
readFile(path, top, mutation, initJar, bio_colnames)
}
\arguments{
\item{path}{results path}

\item{top}{number of top ranks}

\item{mutation}{name of mutation}

\item{initJar}{t of f}

\item{bio_colnames}{total cell type names

readFile(path, top, mutation, initJar, bio_colnames)}
}
\description{
This function reads rank_mutation.txt or rank_mutation_top.txt
and return rankdf
}
