% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporFile.R
\name{reportFile}
\alias{reportFile}
\title{reportFile() function}
\usage{
reportFile(path)
}
\arguments{
\item{path}{results path

reportFile(path)}
}
\description{
This function produces a summary of the results.
}
