% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFile.R
\name{writeFile}
\alias{writeFile}
\title{writeFile() function}
\usage{
writeFile(path, top, mutation, initJar, rp, nround, rankdf)
}
\arguments{
\item{path}{results path}

\item{top}{number of top ranks}

\item{mutation}{name of mutation}

\item{initJar}{t of f}

\item{rp}{number of repetitions}

\item{nround}{:number of rounds}

\item{rankdf}{output of readFile()

writeFile(path, top, mutation, initJar, rp, nround, rankdf)}
}
\description{
This function writes ouput_mutation.txt and ouput_mutation_extract.txt
}
